requirements.txt
# Core ML libraries
tensorflow[and-cuda]==2.19.0
tensorflow-datasets==4.9.9
keras==3.11.3

# MLflow for experiment tracking
mlflow==3.4.0

# Data science libraries
numpy==2.1.3
pandas==2.2.3
scikit-learn==1.6.1

# Image processing
Pillow==10.4.0

# Data loading (for HuggingFace datasets option)
datasets==3.1.0

# Utilities
pyyaml==6.0.2
python-dateutil==2.9.0
psutil==6.1.0

# For matplotlib visualization (if needed for debugging)
matplotlib==3.9.2